[#58] move changed events plugin to use-events libs
