+++ pages/index.vue
@@ -38,9 +38,9 @@
 </template>
 
 <script lang="ts">
-import pluralize from "pluralize"; // eslint-disable-line @conarti/feature-sliced/public-api
+// eslint-disable-next-line @conarti/feature-sliced/public-api
+import pluralize from "pluralize";
 import { defineComponent } from "vue";
-import { useNuxtApp } from "#app"; // eslint-disable-line @conarti/feature-sliced/layers-slices
 import { PageHeader, EventCard, PagePlaceholder } from "~/src/widgets/ui";
 import { PAGE_TYPES } from "~/src/shared/constants";
 import { useEvents } from "~/src/shared/lib/use-events";
@@ -77,19 +77,19 @@ export default defineComponent({
       return events.items.value.filter(({ type }) => type === this.type);
     },
     isEventsPaused() {
-      const { $cachedEvents } = useNuxtApp();
+      const { cachedEvents } = useEvents();
 
-      return $cachedEvents.eventsIdsByType.value[this.type]?.length > 0;
+      return cachedEvents.idsByType.value[this.type]?.length > 0;
     },
     visibleEvents() {
       if (!this.isEventsPaused) {
         return this.allEvents;
       }
 
-      const { $cachedEvents } = useNuxtApp();
+      const { cachedEvents } = useEvents();
 
       return this.allEvents.filter(({ uuid }) =>
-        $cachedEvents.eventsIdsByType.value[this.type]?.includes(uuid)
+        cachedEvents.idsByType.value[this.type]?.includes(uuid)
       );
     },
     hiddenEventsCount() {
@@ -110,12 +110,12 @@ export default defineComponent({
       return events.removeByType(this.type as EventType);
     },
     toggleUpdate() {
-      const { $cachedEvents } = useNuxtApp();
+      const { cachedEvents } = useEvents();
 
       if (this.isEventsPaused) {
-        $cachedEvents.runUpdatesByType(this.type);
+        cachedEvents.runUpdatesByType(this.type);
       } else {
-        $cachedEvents.stopUpdatesByType(this.type);
+        cachedEvents.stopUpdatesByType(this.type);
       }
     },
   },
